Fixing tech qa comments
