+++ src/components/WeatherView/WeatherFixture.js
@@ -1,0 +1,89 @@
+export const belfastWeather = {
+  title: "Yahoo! Weather - Belfast, Northern Ireland, GB",
+  forecast: [
+   {
+    "code": "12",
+    "date": "20 Jul 2016",
+    "day": "Wed",
+    "high": "69",
+    "low": "59",
+    "text": "Rain"
+   },
+   {
+    "code": "39",
+    "date": "21 Jul 2016",
+    "day": "Thu",
+    "high": "67",
+    "low": "58",
+    "text": "Scattered Showers"
+   },
+   {
+    "code": "28",
+    "date": "22 Jul 2016",
+    "day": "Fri",
+    "high": "64",
+    "low": "57",
+    "text": "Mostly Cloudy"
+   },
+   {
+    "code": "28",
+    "date": "23 Jul 2016",
+    "day": "Sat",
+    "high": "66",
+    "low": "54",
+    "text": "Mostly Cloudy"
+   },
+   {
+    "code": "28",
+    "date": "24 Jul 2016",
+    "day": "Sun",
+    "high": "66",
+    "low": "57",
+    "text": "Mostly Cloudy"
+   },
+   {
+    "code": "12",
+    "date": "25 Jul 2016",
+    "day": "Mon",
+    "high": "65",
+    "low": "55",
+    "text": "Rain"
+   },
+   {
+    "code": "39",
+    "date": "26 Jul 2016",
+    "day": "Tue",
+    "high": "64",
+    "low": "54",
+    "text": "Scattered Showers"
+   },
+   {
+    "code": "28",
+    "date": "27 Jul 2016",
+    "day": "Wed",
+    "high": "64",
+    "low": "55",
+    "text": "Mostly Cloudy"
+   },
+   {
+    "code": "12",
+    "date": "28 Jul 2016",
+    "day": "Thu",
+    "high": "63",
+    "low": "53",
+    "text": "Rain"
+   },
+   {
+    "code": "12",
+    "date": "29 Jul 2016",
+    "day": "Fri",
+    "high": "65",
+    "low": "56",
+    "text": "Rain"
+   }
+  ],
+  temperature: "40",
+  todaysCondition: "Cloudy",
+  todaysDate: "Wed, 20 Jul 2016 11:00 AM BST",
+  unit: "F"
+};
