+++ src/components/WeatherView/WeatherStories.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { storiesOf } from '@kadira/storybook';
 import WeatherView from './WeatherView.js';
-import { belfastWeather } from '../../routes/Weather/modules/WeatherHelper.js';
+import { belfastWeather } from './WeatherFixture.js';
 
 const weather = {
-  "weather":  belfastWeather
+  weather:  belfastWeather
 };
 
 storiesOf('Weather', module)
