+++ src/containers/WeatherContainer/WeatherDucks.js
@@ -1,6 +1,6 @@
 import ApiUtils from '../../utils/ApiUtils.js';
 import { weatherbaseUrl } from '../../utils/ApiUtils.js';
-import { belfastWeather } from './WeatherHelper.js';
+import { belfastWeather } from '../../components/WeatherView/WeatherFixture.js';
 
 const query = '?q=select%20*%20from%20weather.forecast%20where%20woeid%20in%20(select%20woeid%20from%20geo.places(1)%20where%20text%3D"london%2C%20uk")&format=json&env=store%3A%2F%2Fdatatables.org%2Falltableswithkeys.json';
 
@@ -15,10 +15,25 @@ const initialState = {
 
 // Reducers
 const REDUCERS = {
-  [FORECAST_SUCCESS]: (state, action) => ({
-    ...state,
-    weather: action.data
-  }),
+  [FORECAST_SUCCESS]: (state, action) => {
+    const { channel } = action.data.query.results;
+    const { title } = channel;
+    const { forecast } = channel.item;
+    const { temp, text, date } = channel.item.condition;
+    const { temperature } = channel.units.temperature;
+
+    return {
+      ...state,
+      weather: {
+        title: title,
+        forecast: forecast,
+        temperature: temp ,
+        todaysCondition: text,
+        todaysDate: date,
+        unit: temperature
+      }
+    };
+  },
   [FORECAST_ERROR]: (state, action) => ({
     ...state,
     error: action.error
@@ -40,7 +55,6 @@ export function forecast () {
 }
 
 function forecastSuccess (result) {
-  console.log(result);
   return {
     type: FORECAST_SUCCESS,
     data: result
